fix: now common OG image functions are in a utility file
